refactor: 찜에서만 쓰이는 앱바 이름 변경
