+++ frontend/src/styles/AppbarSmall.js
